LayoutTests/inspector/debugger/async-stack-trace-expected.txt

CONSOLE MESSAGE: line 83: Unable to post message to http://example.com. Recipient has origin .

Tests for async stack traces.



== Running test suite: AsyncStackTrace
-- Running test case: CheckAsyncStackTrace.RequestAnimationFrame
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
ASYNC CALL STACK:
1: --- requestAnimationFrame ---
2: [F] testRequestAnimationFrame
3: [P] Global Code

-- Running test case: CheckAsyncStackTrace.SetTimeout
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
ASYNC CALL STACK:
1: --- setTimeout ---
2: [F] testSetTimeout
3: [P] Global Code

-- Running test case: CheckAsyncStackTrace.SetInterval
PAUSE #1
CALL STACK:
0: [F] intervalFired
ASYNC CALL STACK:
1: --- setInterval ---
2: [F] testSetInterval
3: [P] Global Code
PAUSE #2
CALL STACK:
0: [F] intervalFired
ASYNC CALL STACK:
1: --- setInterval ---
2: [F] testSetInterval
3: [P] Global Code
PAUSE #3
CALL STACK:
0: [F] intervalFired
ASYNC CALL STACK:
1: --- setInterval ---
2: [F] testSetInterval
3: [P] Global Code

-- Running test case: CheckAsyncStackTrace.ChainedRequestAnimationFrame
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
ASYNC CALL STACK:
1: --- requestAnimationFrame ---
2: [F] testRequestAnimationFrame
3: --- requestAnimationFrame ---
4: [F] testChainedRequestAnimationFrame
5: [P] Global Code

-- Running test case: CheckAsyncStackTrace.ReferenceCounting
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
ASYNC CALL STACK:
1: --- setTimeout ---
2: [F] intervalFired
3: --- setInterval ---
4: [F] testReferenceCounting
5: [P] Global Code

-- Running test case: CheckAsyncStackTrace.AddEventListener
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
1: [F] clickFired
2: [F] testAddEventListener
3: [P] Global Code
ASYNC CALL STACK:
4: --- addEventListener ---
5: [F] testAddEventListener
6: [P] Global Code

-- Running test case: CheckAsyncStackTrace.AddAttributeEventListener
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
1: [F] handleClick
2: [F] testAttributeEventListener
3: [P] Global Code
4: [F] testAttributeEventListener
5: [P] Global Code

-- Running test case: CheckAsyncStackTrace.ReplaceAttributeEventListener
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
1: [F] handleClick2
2: [F] testReplaceAttributeEventListener
3: [P] Global Code
4: [F] testReplaceAttributeEventListener
5: [P] Global Code

-- Running test case: CheckAsyncStackTrace.PostMessage
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
1: [F] postMessageFired
ASYNC CALL STACK:
2: --- postMessage ---
3: [F] messageReceived
4: --- postMessage ---
5: [F] testPostMessage
6: [P] Global Code

-- Running test case: ShouldNotPauseForFailedPostMessage
PASS: Should not pause for failed post message.
-- Running test setup.
Save DebuggerManager.asyncStackTraceDepth

-- Running test case: AsyncStackTrace.DisableStackTrace
PASS: Async stack trace should be null.
-- Running test teardown.
Restore DebuggerManager.asyncStackTraceDepth
-- Running test setup.
Save DebuggerManager.asyncStackTraceDepth

-- Running test case: AsyncStackTrace.ForceTruncationOnCallStackBoundary
Set Debugger.asyncStackTraceDepth equal to 4
PASS: Non-root StackTrace should not be truncated.
PASS: Non-root StackTrace should not be truncated.
PASS: StackTrace root should be truncated.
PASS: StackTrace should be truncated to the nearest call stack.
CALL STACK:
0: [F] pauseThenFinishTest
1: [F] repeat
ASYNC CALL STACK:
2: --- requestAnimationFrame ---
3: [F] repeat
4: --- requestAnimationFrame ---
5: [F] repeat
(remaining call frames truncated)
-- Running test teardown.
Restore DebuggerManager.asyncStackTraceDepth
-- Running test setup.
Save DebuggerManager.asyncStackTraceDepth

-- Running test case: AsyncStackTrace.ForceTruncationWithinCallStack
Set Debugger.asyncStackTraceDepth equal to 5
PASS: Non-root StackTrace should not be truncated.
PASS: Non-root StackTrace should not be truncated.
PASS: Non-root StackTrace should not be truncated.
PASS: StackTrace root should be truncated.
PASS: StackTrace should be truncated to the nearest call stack.
CALL STACK:
0: [F] pauseThenFinishTest
1: [F] repeat
ASYNC CALL STACK:
2: --- requestAnimationFrame ---
3: [F] repeat
4: --- requestAnimationFrame ---
5: [F] repeat
6: --- requestAnimationFrame ---
7: [F] repeat
(remaining call frames truncated)
-- Running test teardown.
Restore DebuggerManager.asyncStackTraceDepth
